Refactor OgTagPreview files into 'ogtag' package
Moved OgTagPreviewScreen and OgTagPreviewViewModel into a dedicated 'ogtag' package for better organization. Updated package declarations accordingly.
